pyproject.toml
[project]
name = "fake-data-agents"
version = "0.1.0"
description = "Python library for generating synthetic data with LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
    "openai==0.27",
    "protobuf>=5.28.2",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["fake_data_agents"]

